Modify location-plugin to allow picking locations on the map
