+++ .changeset/giant-geese-crash.md
@@ -1,0 +1,5 @@
+---
+"@lblod/ember-rdfa-editor-lblod-plugins": minor
+---
+
+Add ability to select geographical positions from a map to the location plugin
