+++ addon/components/location-plugin/nodeview.ts
@@ -2,48 +2,44 @@ import Component from '@glimmer/component';
 import { service } from '@ember/service';
 import IntlService from 'ember-intl/services/intl';
 import { PencilIcon } from '@appuniversum/ember-appuniversum/components/icons/pencil';
+import { type AuPillSignature } from '@appuniversum/ember-appuniversum/components/au-pill';
 
 import { PNode, SayController } from '@lblod/ember-rdfa-editor';
 import { Address } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/location-plugin/utils/address-helpers';
 import { getOutgoingTriple } from '@lblod/ember-rdfa-editor-lblod-plugins/utils/namespace';
 import { EXT } from '@lblod/ember-rdfa-editor-lblod-plugins/utils/constants';
 
-type Args = {
-  getPos: () => number | undefined;
-  node: PNode;
-  controller: SayController;
+type Signature = {
+  Args: {
+    selectNode: () => void;
+    getPos: () => number | undefined;
+    node: PNode;
+    controller: SayController;
+  };
+  Element: AuPillSignature['Element'];
 };
 
-export default class AddressNodeviewComponent extends Component<Args> {
+export default class LocationNodeviewComponent extends Component<Signature> {
   PencilIcon = PencilIcon;
 
   @service declare intl: IntlService;
 
-  get controller() {
-    return this.args.controller;
-  }
-
-  get documentLanguage() {
-    return this.controller.documentLanguage;
-  }
-
   get translations() {
     return {
       placeholder: this.intl.t('editor-plugins.address.nodeview.placeholder', {
-        locale: this.documentLanguage,
+        locale: this.args.controller.documentLanguage,
       }),
     };
   }
 
-  get node() {
-    return this.args.node;
-  }
-
   get address() {
-    return this.node.attrs.value as Address | null;
+    return this.args.node.attrs.value as Address | null;
   }
 
   get label() {
-    return getOutgoingTriple(this.node.attrs, EXT('label'))?.object.value;
+    return (
+      this.address &&
+      getOutgoingTriple(this.args.node.attrs, EXT('label'))?.object.value
+    );
   }
 }
