+++ translations/en-US.yaml
@@ -428,9 +428,45 @@ location-plugin:
     insert: Insert location
     edit: Edit location
     confirm: Set location
+  nodeview:
+    placeholder: Insert location
+  types:
+    address: Address
+    place: Place
   label: Label
   labelPlaceholder: Type something...
   default-label: location
+  search:
+    type: Type
+    place-name:
+      label: Place
+      placeholder: Place name
+    municipality:
+      label: Municipality
+      placeholder: Municipality
+      search-message: Type to search for a municipality
+      no-results: No municipalities found
+    street:
+      label: Street
+      placeholder: Street
+      search-message: Type to search for a street
+      no-results: No streets found
+    housenumber:
+      label: Housenumber
+      placeholder: Housenumber
+    busnumber:
+      label: Busnumber
+      placeholder: Busnumber
+    loading: Resolving address.
+    success:
+      address-found: Address found.
+    errors:
+      address-not-found: We could not resolve the provided address
+      address-not-found-short: Address not found
+      alternative-address: 'Provide another input or proceed with the following address: {address}'
+      http-error: 'Something went wrong while resolving this address, status code: {status}.'
+      projection-error: 'Unable to convert location coordinates: {coords}'
+      contact: In case of persisting issues, contact <a href="mailto:{email}">{email}</a>.
 
 lpdc-plugin:
   insert:
